requirements.txt
mido==1.3.3
packaging==24.2
Pypubsub==4.0.3
python-rtmidi==1.5.8
six==1.16.0
wxPython==4.2.2
